# sandaga-backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sandaga App
  description: >
    'This is a ***Sandaga app API***.'
  termsOfService: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  contact:
    email: info@sandaga.com
  license:
    name: MIT
    url: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: https://sandaga.swagger.io/v2
...
paths:
...
  /api/v1/ads:
    get:
      tags:
        - ads
      summary: Returns all ads
      description: Returns all ads, else empty collection
      operationId: getAllAds
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ads'
    post:
      tags:
        - ads
      summary: Add a new ads
      description: Creates a new ads
      operationId: addAds
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdsReq'
      responses:
        201:
          description: For successful user creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
  /api/v1/ads/{id}:
    get:
      tags:
        - ads
      summary: Returns an ads
      description: Returns ads based on given ads ID.
      operationId: getAdsById
      parameters:
        - name: id
          in: path
          description: ads Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
    patch:
      tags:
        - ads
      summary: Replace a ads
      description: Replace a ads with given id.
      operationId: replacePartialAdsById
      parameters:
        - name: id
          in: path
          description: ads Identifier
          required: true
          schema:
            type: string
      requestBody:
        description: Ads object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ads'
      responses:
        204:
          description: Ads update successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
        404:
          description: Given category ID doesn't exist
          content: { }
    put:
      tags:
        - ads
      summary: Replace a ads
      description: Replace a ads with given id.
      operationId: replaceAdsById
      parameters:
        - name: id
          in: path
          description: ads Identifier
          required: true
          schema:
            type: string
      requestBody:
        description: Ads object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ads'
      responses:
        204:
          description: Ads update successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
        404:
          description: Given category ID doesn't exist
          content: { }
    delete:
      tags:
        - ads
      summary: Deletes ads
      description: Deletes ads based on given ads ID.
      operationId: deleteAdsById
      parameters:
        - name: id
          in: path
          description: ads Identifier
          required: true
          schema:
            type: string
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: { }
  /adv/{id}:
    get:
      tags:
        - ads
      summary: Returns an public ad
      description: Returns public ad based on given ad ID.
      operationId: getPublicAdsById
      parameters:
        - name: id
          in: path
          description: ad Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ads'
...
  /api/v1/addresses:
    get:
      tags:
        - address
      summary: Returns all addresses
      description: Returns all  addresses, else empty collection
      operationId: getAllAddresses
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags:
        - address
      summary: Creates a new user addresses
      description: Creates a new user addresses. Does nothing if address already exists.
      operationId: createAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAddressReq'
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /api/v1/addresses/{id}:
    get:
      tags:
        - address
      summary: Returns an address
      description: Returns an address based on given address ID.
      operationId: getAddressesById
      parameters:
        - name: id
          in: path
          description: address Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    put:
      tags:
        - address
      summary: Replace a Address
      description: Replace an address with given id.
      operationId: replaceAddressById
      parameters:
        - name: id
          in: path
          description: address Identifier
          required: true
          schema:
            type: string
      requestBody:
        description: Address object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAddressReq'
      responses:
        204:
          description: Address update successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        404:
          description: Given address ID doesn't exist
          content: { }
    delete:
      tags:
        - address
      summary: Deletes an address
      description: Deletes an address based on given address ID.
      operationId: deleteAddressesById
      parameters:
        - name: id
          in: path
          description: address Identifier
          required: true
          schema:
            type: string
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: { }
...
components:
  schemas:
...
    Ads:
      type: object
      properties:
        id:
          type: string
        ad_type:
          type: string
        subject:
          type: string
        type:
          type: string
        body:
          type: string
        donation:
          type: boolean
        price_cents:
          type: integer
          format: int32
        price_reco:
          type: integer
          format: int32
        shipping_type:
          type: string
        shipping_cost:
          type: string
        email:
          type: string
        phone:
          type: string
        phone_hidden_information_text:
          type: boolean
        item_condition:
          type: string
        video_game_type:
          type: string
        console_model:
          type: string
        console_brand:
          type: string
        image_sound_type_of_product:
          type: string
        image_sound_product:
          type: string
        phone_brand:
          type: string
        phone_color:
          type: string
        phone_model:
          type: string
        phone_memory:
          type: string
        furniture_type:
          type: string
        furniture_material:
          type: string
        furniture_color:
          type: string
        home_appliance_type:
          type: string
        home_appliance_product:
          type: string
        home_appliance_brand:
          type: string
        table_art_product:
          type: string
        table_art_material:
          type: string
        decoration_type:
          type: string
        linens_type:
          type: string
        linens_product:
          type: string
        linens_material:
          type: string
        diy_type:
          type: string
        diy_product:
          type: string
        gardering_type:
          type: string
        gardering_product:
          type: string
        clothing_type:
          type: string
        clothing_st:
          type: string
        clothing_category:
          type: string
        clothing_brand:
          type: string
        clothing_color:
          type: string
        clothing_condition:
          type: string
        shoe_type:
          type: string
        shoe_category:
          type: string
        shoe_size:
          type: string
        shoe_brand:
          type: string
        accessories_univers:
          type: string
        accessories_type:
          type: string
        accessories_brand:
          type: string
        accessories_material:
          type: string
        watches_jewels_brand:
          type: string
        watches_jewels_type:
          type: string
        watches_jewels_material:
          type: string
        baby_equipment_type:
          type: string
        baby_equipment_brand:
          type: string
        baby_clothing_category:
          type: string
        baby_clothing_brand:
          type: string
        baby_age:
          type: string
        bicycle_type:
          type: string
        bicycle_size:
          type: string
        bicycode:
          type: string
        sports_hobbies_activity:
          type: string
        toy_age:
          type: string
        toy_type:
          type: string
        matpro_agriculture_equipment_type:
          type: string
        regdate:
          type: string
      xml:
        name: Ads
    AdsReq:
      allOf:
        - $ref: '#/components/schemas/Ads'
        - type: object
          properties:
            userId:
              type: string
            categoryId:
              type: string
            addressId:
              type: string

      xml:
        name: AddAddressReq
...
    Address:
      type: object
      properties:
        id:
          description: Address identifier
          type: string
        number:
          description: house of flat number
          type: string
        street:
          description: street name
          type: string
        city:
          description: city name
          type: string
        state:
          description: state name
          type: string
        country:
          description: country name
          type: string
        pincode:
          description: postal code
          type: string
      xml:
        name: Address
    AddAddressReq:
      allOf:
        - $ref: '#/components/schemas/Address'
        - type: object
      xml:
        name: AddAddressReq
...